now using explicite get_final_layer_norm implementation
